+++ Sources/SwiftPackageScripts/SwiftPackageScripts.swift
@@ -1,2 +1,0 @@
-/// This class is here to verify that all scripts work.
-public class SampleClass {}