+++ Tests/SwiftPackageScriptsTests/SwiftPackageBuildScriptsTests.swift
@@ -1,6 +1,0 @@
-import Testing
-import SwiftPackageScripts
-
-@Test func example() async throws {
-    #expect(1 == 1)
-}